Refactor :scroll left/right --open-selected
Instead of having the working directory handler directly load the
directory, we now defer this to the LibraryModel. This way, the model
can decide if the path selected in the new directory is an image. If
yes, load it - with this image focused.

Thus, we don't end up jumping positions like previously, when we enter a
directory that:
* Has images
* Has an image in the stored position
Where:
1) We would load the directory into the library => paths visible in
   library
2) We would load the stored position => new path selected
3) We would load all images into the image filelist
4) We would open the first image
5) The library would select the first image => jump to new path which is
   not desirable